python/pyproject.toml
[project]
name = "cast-workers"
version = "0.1.0"
description = "Python ML workers for Cast"
requires-python = ">=3.11"
dependencies = [
    "openai-whisper>=20231117",
    "open-clip-torch>=2.24.0",
    "insightface>=0.7.3",
    "onnxruntime>=1.17.0",
    "opencv-python-headless>=4.9.0",
    "torch>=2.2.0",
    "pillow>=10.2.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
